requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
python-dotenv